Add files via upload
The Luhn algorithm validator assists in checking and distinguishing the valid numbers from the mistyped or incorrect inputs. It is a method widely used by government and corporate entities to eliminate accidental errors by customers or users while using credit cards or identification numbers for online payments or transactions.
The Luhn algorithm formula for verification of identification numbers includes the following steps,

Step 1 –  The sequence of numbers starting from the first number on the right or the rightmost number is multiplied by two or doubled, and this follows for every second number from the first rightmost number. 
Step 2 –  If the double of a digit is equal to or larger than 10, then the sum of the resultant double-digit numbers is taken,
Step 3 – Next step is to calculate the total of all the resultants that we have after multiplying by 2,
Step 4 –  After realizing the sum of all digits, the check key or control digit ‘x’ equals = x = (10 – (x mod 10) mod 10). Thus, the value of ‘x’ is the checksum or check key according to Luhn Algorithm to verify a given sequence of numbers. Usually, the ‘x’ is the last digit of the identification number sequence that the algorithm determines.